Remove lib/web dependency from tsh
Refactors the terminal stream code in lib/web into a subpackage
so that tsh no longer depends on the entire lib/web package.
