+++ build.assets/genproto.sh
@@ -47,7 +47,7 @@ main() {
     --path=api/proto/teleport/attestation/ \
     --path=api/proto/teleport/usageevents/ \
     --path=api/proto/teleport/mfa/ \
-    --path=proto/teleport/lib/web/envelope.proto \
+    --path=proto/teleport/lib/web/terminal/envelope.proto \
     --exclude-path=api/proto/teleport/legacy/client/proto/event.proto
   cp -r gogogen/github.com/gravitational/teleport/. .
   # error out if there's anything outside of github.com/gravitational/teleport
@@ -60,7 +60,7 @@ main() {
     --exclude-path=api/proto/teleport/attestation/ \
     --exclude-path=api/proto/teleport/usageevents/ \
     --exclude-path=api/proto/teleport/mfa/ \
-    --exclude-path=proto/teleport/lib/web/envelope.proto \
+    --exclude-path=proto/teleport/lib/web/terminal/envelope.proto \
     --exclude-path=proto/prehog/
 
   # Generate event.proto separately because we only want to run it on this
